database/requirements.txt
psycopg
python-dotenv
